store the codes from G4 directly
Previously we were expecting to translate the 'process' and 'subprocess' codes from GEANT4 into our own enumeration.  This is laborious and error-prone, however, and in the first main use case (ND), we're supplied the GEANT4 codes directly.  I don't believe the FD CAFMaker fills this branch yet.
So we'll store the codes directly from GEANT4, and include links to the documentation where the values are explained.  This is probably the most maintainable solution anyway.